tide::data::config::serialization::color_serde_maybe
Function deserialize

pub fn deserialize<'de, D>(deserializer: D) -> Result<Option<Color>, D::Error>
where
    D: Deserializer<'de>,

Deserializes an optional hexadecimal string into an Option<Color>.

Returns None if the field is missing or the hex is invalid.

Errors

Returns a SerDe deserialization error if the input is not a valid string.