tide::data::style
Module file_tree

Functions

main_style

Returns the default text color style for the main file in the file tree.

selected_file

Returns the style for a container representing a selected file in the file tree.

svg_icon

Returns the color styling for SVG icons.