tide::editor::bindings
Function bindings

pub fn bindings(key_press: KeyPress) -> Option<Binding<Message>>

Maps key presses to editor bindings. Returns an optional [Binding] corresponding to the key combination to a editing::Message.

This includes:

Tab –> Add four spaces

Ctrl + S –> Force preview and save the current file

Ctrl + Arrow Right –> Move to the right boundary of a word

Ctrl + Shift + Arrow Right –> Move to the right boundary of a word and select all of its characters

Ctrl + Arrow Left –> Move to the left boundary of a word

Ctrl + Shift + Arrow Left –> Move to the left boundary of a word and select all of its characters

Del –> Delete the next character

Ctrl + O –> Move to the end of the line and break the current line

Ctrl + E –> Export current project as a PDF

Ctrl + Space –> Open the autocomplete context