tide::file_manager::import::load
Function load_source_file

fn load_source_file(
    path: &PathBuf,
    root: &PathBuf,
) -> Result<ImportedFile, Error>

Loads a .typ Typst source file and wraps it as an ImportedFile::TypstSource instance.

Errors

Returns an Error if the file cannot be read or the virtual path could not be computed.