Struct tide::http::headers::HeaderValues

pub struct HeaderValues { /* fields omitted */ }

A list of HeaderValues.

This always contains at least one header value.

Implementations

impl HeaderValues

pub fn append(&mut self, other: &mut HeaderValues)

Move all values from other into self, leaving other empty.

pub fn get(&self, index: usize) -> Option<&HeaderValue>

Returns a reference or a value depending on the type of index.

pub fn get_mut(&mut self, index: usize) -> Option<&mut HeaderValue>

Returns a mutable reference or a value depending on the type of index.

pub fn contains(&self, value: &HeaderValue) -> bool

Returns true if there is a value corresponding to the specified HeaderValue in the list, false otherwise.

pub fn last(&self) -> &HeaderValue

Returns the last HeaderValue.

pub fn iter(&self) -> Values<'_>

An iterator visiting all header values in arbitrary order.