Struct tide::listener::FailoverListener

pub struct FailoverListener<State> { /* fields omitted */ }

FailoverListener allows tide to attempt to listen in a sequential order to any number of ports/addresses. The first successful listener is used.

Example:

fn main() -> Result<(), std::io::Error> {
   async_std::task::block_on(async {
       tide::log::start();
       let mut app = tide::new();
       app.at("/").get(|_| async { Ok("Hello, world!") });

       let mut listener = tide::listener::FailoverListener::new();
       listener.add("127.0.0.1:8000")?;
       listener.add(async_std::net::TcpListener::bind("127.0.0.1:8001").await?)?;
       listener.add("http+unix://unix.socket")?;
   
       app.listen(listener).await?;
       Ok(())
   })
}

Implementations

impl<State> FailoverListener<State> where
State: Clone + Send + Sync + 'static,

pub fn new() -> Self

creates a new FailoverListener

pub fn add<L>(&mut self, listener: L) -> Result<()> where
L: ToListener<State>,

Adds any ToListener to this FailoverListener. An error result represents a failure to convert the ToListener into a Listener.

let mut listener = tide::listener::FailoverListener::new();
listener.add("127.0.0.1:8000")?;
listener.add(("localhost", 8001))?;
listener.add("http+unix:///var/run/tide")?;

pub fn with_listener<L>(self, listener: L) -> Self where
L: ToListener<State>,

FailoverListener::with_listener allows for chained construction of a FailoverListener:

app.listen(
    FailoverListener::new()
        .with_listener("127.0.0.1:8080")
        .with_listener(("localhost", 8081)),
).await?;