Module tide::prelude

The Tide prelude.

Re-exports

pub use crate::listener::Listener;

Macros

json

Construct a serde_json::Value from a JSON literal.

Traits

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Status

Provides the status method for Result and Option.

Derive Macros

Deserialize

Serialize