tide::screen::component::file_tree
Struct FileTree

pub struct FileTree {
    file_system: Dir,
    main_path: Option<PathBuf>,
    pub selected_path: Option<PathBuf>,
    clicked_path: Option<PathBuf>,
}

Represents the interactive file tree UI component of Tide.

Stores and manages state related to the current directory structure, the selected and clicked files, and the main Typst file.

Fields
file_system: Dir

Internal representation of the file system.

main_path: Option<PathBuf>

Path to the file marked as the main Typst file.

selected_path: Option<PathBuf>

Path to the currently selected file in the UI.

clicked_path: Option<PathBuf>

Path to the file that was most recently clicked.

Implementations

impl FileTree

pub fn new(root_path: PathBuf) -> Self

Creates a new FileTree rooted at the given directory path.

pub fn view(&self) -> (Element<'_, Message>, f32)

Builds and returns the file tree view as an [Element].

Returns the UI element and its height.

pub fn add_file(&mut self, path: &PathBuf)

Adds a file to the file tree.

The file is inserted into the internal list of files and sorted alphabetically.

pub fn refresh(&mut self)

Refreshes file entries in the tree.

Useful after a file has been deleted or modified externally.

pub fn fold(&mut self, path: &PathBuf)

Toggles open/closed state of a directory node.

pub fn change_main(&mut self, path: &PathBuf)

Changes the main Typst file path.

If the path is already the main file, no change is made.

pub fn change_selected(&mut self, path: &PathBuf)

Changes the currently selected file.

pub fn change_clicked(&mut self, path: Option<PathBuf>)

Changes the last clicked file (used for file actions).