tide::screen
Module welcome

Structs

Welcome

Represents the main Welcome page shown on Tide startup.

Enums

Message

Messages handled by the Welcome view.

Constants

FILE_BUTTON_HEIGHT 🔒

WELCOME_BUTTON_HEIGHT 🔒

WELCOME_BUTTON_PADDING 🔒

WELCOME_BUTTON_WIDTH 🔒

Statics

LOGO

Functions

file_button 🔒

Creates a clickable button representing a recent file or project.

welcome_button 🔒

Creates a (big) button for the welcome screen with standard style.