Trait tide::sessions::SessionStore

pub trait SessionStore: 'static + Clone + Send + Sync + Debug {
#[must_use]    pub fn load_session<'life0, 'async_trait>(
        &'life0 self, 
        cookie_value: String
    ) -> Pin<Box<dyn Future<Output = Result<Option<Session>, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    pub fn store_session<'life0, 'async_trait>(
        &'life0 self, 
        session: Session
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    pub fn destroy_session<'life0, 'async_trait>(
        &'life0 self, 
        session: Session
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    pub fn clear_store<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

An async session backend.

Required methods

#[must_use]pub fn load_session<'life0, 'async_trait>(
&'life0 self,
cookie_value: String
) -> Pin<Box<dyn Future<Output = Result<Option<Session>, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Get a session from the storage backend.

The input is expected to be the value of an identifying cookie. This will then be parsed by the session middleware into a session if possible

#[must_use]pub fn store_session<'life0, 'async_trait>(
&'life0 self,
session: Session
) -> Pin<Box<dyn Future<Output = Result<Option<String>, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Store a session on the storage backend.

The return value is the value of the cookie to store for the user that represents this session

#[must_use]pub fn destroy_session<'life0, 'async_trait>(
&'life0 self,
session: Session
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Remove a session from the session store

#[must_use]pub fn clear_store<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Empties the entire store, destroying all sessions

Loading content...