Struct tide::Route

pub struct Route<'a, State> { /* fields omitted */ }

A handle to a route.

All HTTP requests are made against resources. After using Server::at (or Route::at) to establish a route, the Route type can be used to establish endpoints for various HTTP methods at that path. Also, using nest, it can be used to set up a subrouter.

Implementations

impl<'a, State: Clone + Send + Sync + 'static> Route<'a, State>

pub fn at<'b>(&'b mut self, path: &str) -> Route<'b, State>

Extend the route with the given path.

pub fn path(&self) -> &str

Get the current path.

pub fn with<M>(&mut self, middleware: M) -> &mut Self where
M: Middleware<State>,

Apply the given middleware to the current route.

pub fn reset_middleware(&mut self) -> &mut Self

Reset the middleware chain for the current route, if any.

pub fn nest<InnerState>(&mut self, service: Server<InnerState>) -> &mut Self where
State: Clone + Send + Sync + 'static,
InnerState: Clone + Send + Sync + 'static,

Nest a Server at the current path.

pub fn serve_dir(&mut self, dir: impl AsRef<Path>) -> Result<()>

Serve a directory statically.

Each file will be streamed from disk, and a mime type will be determined based on magic bytes.

Examples

Serve the contents of the local directory ./public/images/* from localhost:8080/images/*.

#[async_std::main]
async fn main() -> Result<(), std::io::Error> {
    let mut app = tide::new();
    app.at("/public/images").serve_dir("images/")?;
    app.listen("127.0.0.1:8080").await?;
    Ok(())
}

pub fn method(&mut self, method: Method, ep: impl Endpoint<State>) -> &mut Self

Add an endpoint for the given HTTP method

pub fn all(&mut self, ep: impl Endpoint<State>) -> &mut Self

Add an endpoint for all HTTP methods, as a fallback.

Routes with specific HTTP methods will be tried first.

pub fn get(&mut self, ep: impl Endpoint<State>) -> &mut Self

Add an endpoint for GET requests

pub fn head(&mut self, ep: impl Endpoint<State>) -> &mut Self

Add an endpoint for HEAD requests

pub fn put(&mut self, ep: impl Endpoint<State>) -> &mut Self

Add an endpoint for PUT requests

pub fn post(&mut self, ep: impl Endpoint<State>) -> &mut Self

Add an endpoint for POST requests

pub fn delete(&mut self, ep: impl Endpoint<State>) -> &mut Self

Add an endpoint for DELETE requests

pub fn options(&mut self, ep: impl Endpoint<State>) -> &mut Self

Add an endpoint for OPTIONS requests

pub fn connect(&mut self, ep: impl Endpoint<State>) -> &mut Self

Add an endpoint for CONNECT requests

pub fn patch(&mut self, ep: impl Endpoint<State>) -> &mut Self

Add an endpoint for PATCH requests

pub fn trace(&mut self, ep: impl Endpoint<State>) -> &mut Self

Add an endpoint for TRACE requests