Trait tide::Status

pub trait Status<T, E>: Sealed {
    fn status<S>(self, status: S) -> Result<T, Error>
    where
        S: TryInto<StatusCode>,
        <S as TryInto<StatusCode>>::Error: Debug;

    fn with_status<S, F>(self, f: F) -> Result<T, Error>
    where
        S: TryInto<StatusCode>,
        F: FnOnce() -> S,
        <S as TryInto<StatusCode>>::Error: Debug;
}

Provides the status method for Result and Option.

This trait is sealed and cannot be implemented outside of http-types.

Required methods

fn status<S>(self, status: S) -> Result<T, Error> where
S: TryInto<StatusCode>,
<S as TryInto<StatusCode>>::Error: Debug,

Wrap the error value with an additional status code.

fn with_status<S, F>(self, f: F) -> Result<T, Error> where
S: TryInto<StatusCode>,
F: FnOnce() -> S,
<S as TryInto<StatusCode>>::Error: Debug,

Wrap the error value with an additional status code that is evaluated lazily only once an error does occur.