Struct tide::utils::After

pub struct After<F>(pub F);

Define a middleware that operates on outgoing responses.

This middleware is useful because it is not possible in Rust yet to use closures to define inline middleware.

Examples

use tide::{utils, http, Response};

let mut app = tide::new();
app.with(utils::After(|res: Response| async move {
    match res.status() {
        http::StatusCode::NotFound => Ok("Page not found".into()),
        http::StatusCode::InternalServerError => Ok("Something went wrong".into()),
        _ => Ok(res),
    }
}));