Module tide::log

Event logging types.

Examples

use tide::log;

log::start();

log::info!("Hello cats");
log::debug!("{} wants tuna", "Nori");
log::error!("We're out of tuna!");
log::info!("{} are hungry", "cats", {
    cat_1: "Chashu",
    cat_2: "Nori",
});

Macros

debug
Logs a message at the debug level.

error
Logs a message at the error level.

info
Logs a message at the info level.

log
The standard logging macro.

trace
Logs a message at the trace level.

warn
Logs a message at the warn level.

Structs

LogMiddleware
Log all incoming requests and responses.

Enums

Level
An enum representing the available verbosity levels of the logger.

LevelFilter
An enum representing the available verbosity level filters of the logger.

Functions

max_level
Returns the current maximum log level.

start
Start logging.

with_level
Start logging with a log level.